SECURITY WARNING: stop encrypting with a label
This branch is an experimenting in using KMS for decrypting data. The
KMS support for asymmetric keys is pretty good, but sadly it's missing
**ONE** feature currently used by sealed-secrets: labels.

If we encrypt with a label, we're unable to decrypt the cipher text. With
no label, we can.

sealed-secrets uses labels to ensure a sealed secret can only be
decrypted when it's inside the namespace it was targeted for. If someone
takes a sealed secret and adds it to a different namespace on the same
cluster, it will fail to decrypt - despite the fact the controller is
using the same key.